url's and views cleanup
