+++ base/urls/order_urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+from base.views import order_views as views
+
+
+urlpatterns = [
+
+]
