+++ base/urls/product_urls.py
@@ -1,0 +1,16 @@
+
+# from re import U
+# from urllib.parse import urljoin
+from django.urls import path
+from base.views import product_views as views
+
+
+urlpatterns = [
+
+    path('', views.getProducts , name='products'),
+    path('<pk>/', views.getProduct, name='product'),  #<str:pk>
+]
+
+ 
+
+
